requirements.txt
altair==5.0.1
area==1.1.1
attrs==23.1.0
blinker==1.6.2
branca==0.3.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
Fiona==1.9.4.post1
folium==0.14.0
fonttools==4.40.0
geojson==2.5.0
geopandas==0.13.2
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
numpy==1.25.0
packaging==23.1
pandas==2.0.3
patsy==0.5.3
Pillow==9.5.0
plotly==5.15.0
plotly-express==0.4.1
protobuf==4.23.3
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyowm==3.3.0
pyparsing==3.1.0
pyproj==3.6.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyweatherbit==2.1.1
PyYAML==6.0
requests==2.31.0
responses==0.23.1
rich==13.4.2
scipy==1.11.1
shapely==2.0.1
six==1.16.0
smmap==5.0.0
statsmodels==0.14.0
streamlit==1.24.0
streamlit-card==0.0.61
streamlit-elements==0.1.0
streamlit-folium==0.12.0
streamlit-lottie==0.0.5
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
types-PyYAML==6.0.12.10
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
